apps/backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.8",
  "sqlalchemy>=2.0",
  "alembic>=1.13",
  "psycopg[binary]>=3.2",        # PostgreSQL (로컬 테스트면 sqlite3로 대체 가능)
  "passlib[bcrypt]>=1.7",
  "python-jose[cryptography]>=3.3",
  "python-multipart>=0.0.9",     # 폼 파서 (선택)
]

[tool.pytest.ini_options]
testpaths = ["apps/backend/tests"]
addopts = "-q"